pyproject.toml
[project]
name = "databricks-study"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel Mizuno",email = "gabrielmizuno@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faker (>=36.1.1,<37.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pendulum (>=3.0.0,<4.0.0)",
    "pyarrow (>=19.0.1,<20.0.0)",
    "google-cloud-pubsub (==2.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3